requirements.txt
astroid==3.3.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
fonttools==4.56.0
imbalanced-learn==0.13.0
isort==6.0.0
joblib==1.2.0
kiwisolver==1.4.8
matplotlib==3.10.0
mccabe==0.7.0
numpy==1.24.3
packaging==24.2
pandas==1.5.3
pillow==11.1.0
platformdirs==4.3.6
pylint==3.3.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.3.2
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.5.0
tomlkit==0.13.2
tzdata==2025.1




